absolute imports, individual cmp folder, proper naming
