+++ web/components/AlbumCover/AlbumCover.module.scss
