+++ web/components/CardEquipment/CardEquipment.module.scss
