+++ web/components/CardNews/CardNews.module.scss
