+++ web/components/CoverImage/CoverImage.module.scss
