+++ web/pages/_app.js
@@ -1,10 +1,10 @@
-import '../styles/globals.css';
-import '../styles/offcanvas.css';
-import '../styles/Header.css';
-import { getClient } from '../lib/sanity.server';
-import { rrssQuery } from '../lib/queries';
+import 'styles/globals.css';
+import 'styles/offcanvas.css';
+import 'styles/Header.css';
+import { getClient } from 'lib/sanity.server';
+import { rrssQuery } from 'lib/queries';
 import { useState } from 'react';
-import { Context } from '../components/context';
+import { Context } from 'components/context';
 
 function MyApp({ Component, pageProps, rrss }) {
     const [context, setContext] = useState(rrss);
