+++ web/pages/categoria/contribuciones.js
@@ -1,12 +1,12 @@
-import { getClient } from '../../lib/sanity.server';
-import { contributionPostQuery } from '../../lib/queries';
-import Layout from '../../components/layout';
-import AlbumCover from '../../components/albumCover';
+import { getClient } from 'lib/sanity.server';
+import { contributionPostQuery } from 'lib/queries';
+import Layout from 'components/Layout/Layout';
+import AlbumCover from 'components/AlbumCover/AlbumCover';
 import { useState } from 'react';
-import Gallery from '../../components/gallery';
-import Card from '../../components/Cards/cardNews';
-import styles from '../../styles/PageSidebar.module.css';
-import NavCategorias from '../../components/Navigation/NavCategorias';
+import Gallery from 'components/Gallery/Gallery';
+import Card from 'components/CardNews/CardNews';
+import styles from 'styles/PageSidebar.module.css';
+import NavCategorias from 'components/NavCategorias/NavCategorias';
 
 export default function ContribucionesCategory({ news, albums }) {
     const [selectedAlbum, setSelectedAlbum] = useState([]);
