+++ web/pages/galeria.js
@@ -1,10 +1,10 @@
 import React, { useState } from 'react';
-import Layout from '../components/layout';
-import { getClient } from '../lib/sanity.server';
-import { albumsQuery } from '../lib/queries';
-import Gallery from '../components/gallery';
-import AlbumCover from '../components/albumCover';
-import HeroPage from '../components/Heros/HeroPage';
+import Layout from 'components/Layout/Layout';
+import { getClient } from 'lib/sanity.server';
+import { albumsQuery } from 'lib/queries';
+import Gallery from 'components/Gallery/Gallery';
+import AlbumCover from 'components/AlbumCover/AlbumCover';
+import HeroPage from 'components/HeroPage/HeroPage';
 import Fade from 'react-reveal/Fade';
 
 export default function Galeria({ albums }) {
