+++ web/pages/historia.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import ArticleContent from '../components/articleContent';
-import { getClient } from '../lib/sanity.server';
-import { pagesQuery } from '../lib/queries';
+import ArticleContent from 'components/ArticleContent/ArticleContent';
+import { getClient } from 'lib/sanity.server';
+import { pagesQuery } from 'lib/queries';
 import { SRLWrapper } from 'simple-react-lightbox';
 
 const options = {
